refactor: 组件文件夹格式统一
