+++ app/assets/pages/api-test-page/index.jsx
@@ -7,10 +7,10 @@ import { getEnvByUrl } from '../../utils/utils';
 
 import Info from '../../components/info/index';
 import Editor from '../../components/editor/index';
-import AuthForm from './components/AuthForm';
-import MyTabs from './components/MyTabs';
+import AuthForm from './components/auth-form/index';
+import MyTabs from './components/my-tabs/index';
 import BulkEditor from '../../components/bulk-editor/index';
-import Result from './components/Result';
+import Result from './components/result/index';
 
 import './index.less';
